♻️ refactor: 펀딩 상세 페이지 하나로 합침
